% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formAdminModule.R
\name{formAdminUI}
\alias{formAdminUI}
\alias{formAdminModule}
\title{Shiny module for form configuration}
\usage{
formAdminUI(id)

formAdminModule(input, output, session, .reg = NULL)
}
\description{
Shiny module for form configuration
}
