% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formAdminModule.R
\name{formAdminUI}
\alias{formAdminUI}
\alias{formAdminModule}
\title{Shiny module for form configuration}
\usage{
formAdminUI(
  id,
  title = "Formulieropstelling",
  title_deleted = "Verwijderde invoervelden",
  tab_deleted_input_fields = TRUE,
  option_layout = TRUE,
  option_add_field = TRUE,
  option_delete_field = TRUE,
  header_ui = NULL
)

formAdminModule(
  input,
  output,
  session,
  .reg = NULL,
  allow_delete_option = FALSE
)
}
\arguments{
\item{id}{Shiny input id}

\item{title}{Title for the tab panel with settings}

\item{title_deleted}{Title for the tab panel with deleted fields}

\item{header_ui}{UI to add at the top of the panel with settings}
}
\description{
Shiny module for form configuration
}
