% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formClass.R
\name{formClass}
\alias{formClass}
\title{R6 Class voor Registratie formulier}
\description{
R6 Class voor Registratie formulier

R6 Class voor Registratie formulier
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{formClass$new()}}
\item \href{#method-connect_to_database}{\code{formClass$connect_to_database()}}
\item \href{#method-close}{\code{formClass$close()}}
\item \href{#method-make_column}{\code{formClass$make_column()}}
\item \href{#method-read_table}{\code{formClass$read_table()}}
\item \href{#method-append_data}{\code{formClass$append_data()}}
\item \href{#method-table_columns}{\code{formClass$table_columns()}}
\item \href{#method-query}{\code{formClass$query()}}
\item \href{#method-has_value}{\code{formClass$has_value()}}
\item \href{#method-replace_value_where}{\code{formClass$replace_value_where()}}
\item \href{#method-get_by_id}{\code{formClass$get_by_id()}}
\item \href{#method-make_choices}{\code{formClass$make_choices()}}
\item \href{#method-choices_from_json}{\code{formClass$choices_from_json()}}
\item \href{#method-rename_definition_table}{\code{formClass$rename_definition_table()}}
\item \href{#method-get_input_fields}{\code{formClass$get_input_fields()}}
\item \href{#method-get_form_fields}{\code{formClass$get_form_fields()}}
\item \href{#method-get_next_formorder_number}{\code{formClass$get_next_formorder_number()}}
\item \href{#method-add_input_field_to_form}{\code{formClass$add_input_field_to_form()}}
\item \href{#method-check_uniqueness_column_name}{\code{formClass$check_uniqueness_column_name()}}
\item \href{#method-edit_label_field}{\code{formClass$edit_label_field()}}
\item \href{#method-from_json}{\code{formClass$from_json()}}
\item \href{#method-to_json}{\code{formClass$to_json()}}
\item \href{#method-edit_options_field}{\code{formClass$edit_options_field()}}
\item \href{#method-is_color}{\code{formClass$is_color()}}
\item \href{#method-edit_options_colors}{\code{formClass$edit_options_colors()}}
\item \href{#method-amend_options_colors}{\code{formClass$amend_options_colors()}}
\item \href{#method-set_options_order}{\code{formClass$set_options_order()}}
\item \href{#method-amend_options_order}{\code{formClass$amend_options_order()}}
\item \href{#method-edit_formulier_setup}{\code{formClass$edit_formulier_setup()}}
\item \href{#method-edit_zichtbaarheid_invoerveld}{\code{formClass$edit_zichtbaarheid_invoerveld()}}
\item \href{#method-edit_verwijder_datum}{\code{formClass$edit_verwijder_datum()}}
\item \href{#method-amend_formfield_order}{\code{formClass$amend_formfield_order()}}
\item \href{#method-reset_volgorde_invoerveld}{\code{formClass$reset_volgorde_invoerveld()}}
\item \href{#method-column_type_from_field_type}{\code{formClass$column_type_from_field_type()}}
\item \href{#method-prepare_data_table}{\code{formClass$prepare_data_table()}}
\item \href{#method-write_new_registration}{\code{formClass$write_new_registration()}}
\item \href{#method-edit_registration}{\code{formClass$edit_registration()}}
\item \href{#method-read_registrations}{\code{formClass$read_registrations()}}
\item \href{#method-get_registration_by_id}{\code{formClass$get_registration_by_id()}}
\item \href{#method-filter_period}{\code{formClass$filter_period()}}
\item \href{#method-read_definition}{\code{formClass$read_definition()}}
\item \href{#method-clone}{\code{formClass$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Make new form object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$new(
  config_file = "conf/config.yml",
  what,
  schema = NULL,
  def_table = "formulier_velden",
  def_columns = list(id_form = "id_formulierveld", column_field = "kolomnaam_veld",
    label_field = "label_veld", type_field = "type_veld", order_field = "volgorde_veld",
    form_section = "formulier_sectie", visible = "zichtbaar", date_deleted =
    "datum_uitgeschakeld", options = "opties", order_options = "volgorde_opties", colors
    = "kleuren", removable = "kan_worden_verwijderd"),
  data_table = "registrations",
  data_columns = list(id = "id_registratie", name = "naam_registratie", time_created =
    "invoerdatum", time_modified = "wijzigdatum", user = "user_id"),
  inject = NULL,
  pool = TRUE,
  sqlite = NULL,
  default_color = "#3333cc"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config_file}}{Path to DB config}

\item{\code{schema}}{DB schema}

\item{\code{pool}}{If TRUE, connects with dbPool}

\item{\code{sqlite}}{If path to an SQLite file, uses SQLite.}

\item{\code{default_color}}{Default color}

\item{\code{wat}}{Entry in config for DB connection}

\item{\code{table}}{Table in DB (in schema) that holds form config}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connect_to_database"></a>}}
\if{latex}{\out{\hypertarget{method-connect_to_database}{}}}
\subsection{Method \code{connect_to_database()}}{
Connect to a database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$connect_to_database(
  config_file = NULL,
  schema = NULL,
  what = NULL,
  pool = TRUE,
  sqlite = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-close"></a>}}
\if{latex}{\out{\hypertarget{method-close}{}}}
\subsection{Method \code{close()}}{
Close database connection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$close()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-make_column"></a>}}
\if{latex}{\out{\hypertarget{method-make_column}{}}}
\subsection{Method \code{make_column()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$make_column(table, column, type = "varchar")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read_table"></a>}}
\if{latex}{\out{\hypertarget{method-read_table}{}}}
\subsection{Method \code{read_table()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$read_table(table, lazy = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-append_data"></a>}}
\if{latex}{\out{\hypertarget{method-append_data}{}}}
\subsection{Method \code{append_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$append_data(table, data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-table_columns"></a>}}
\if{latex}{\out{\hypertarget{method-table_columns}{}}}
\subsection{Method \code{table_columns()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$table_columns(table)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-query"></a>}}
\if{latex}{\out{\hypertarget{method-query}{}}}
\subsection{Method \code{query()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$query(txt, glue = TRUE, quiet = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-has_value"></a>}}
\if{latex}{\out{\hypertarget{method-has_value}{}}}
\subsection{Method \code{has_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$has_value(table, column, value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-replace_value_where"></a>}}
\if{latex}{\out{\hypertarget{method-replace_value_where}{}}}
\subsection{Method \code{replace_value_where()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$replace_value_where(
  table,
  col_replace,
  val_replace,
  col_compare,
  val_compare,
  query_only = FALSE,
  quiet = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_by_id"></a>}}
\if{latex}{\out{\hypertarget{method-get_by_id}{}}}
\subsection{Method \code{get_by_id()}}{
Get a row from the form definition by the form id.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$get_by_id(id_form)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-make_choices"></a>}}
\if{latex}{\out{\hypertarget{method-make_choices}{}}}
\subsection{Method \code{make_choices()}}{
Make choices (for selectInput) based on values and names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$make_choices(
  values_from,
  names_from = values_from,
  data = NULL,
  sort = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-choices_from_json"></a>}}
\if{latex}{\out{\hypertarget{method-choices_from_json}{}}}
\subsection{Method \code{choices_from_json()}}{
Unpack a JSON field to make a named vector
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$choices_from_json(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rename_definition_table"></a>}}
\if{latex}{\out{\hypertarget{method-rename_definition_table}{}}}
\subsection{Method \code{rename_definition_table()}}{
Rename database table to correct internal column names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$rename_definition_table(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_input_fields"></a>}}
\if{latex}{\out{\hypertarget{method-get_input_fields}{}}}
\subsection{Method \code{get_input_fields()}}{
Get all non-deleted form input fields
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$get_input_fields(zichtbaarheid = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_form_fields"></a>}}
\if{latex}{\out{\hypertarget{method-get_form_fields}{}}}
\subsection{Method \code{get_form_fields()}}{
Get form fields for left or right column of the form
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$get_form_fields(form_section)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_next_formorder_number"></a>}}
\if{latex}{\out{\hypertarget{method-get_next_formorder_number}{}}}
\subsection{Method \code{get_next_formorder_number()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$get_next_formorder_number(form_section)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_input_field_to_form"></a>}}
\if{latex}{\out{\hypertarget{method-add_input_field_to_form}{}}}
\subsection{Method \code{add_input_field_to_form()}}{
Adds an input field to a form
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$add_input_field_to_form(label_field, type_field, form_section)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{label_field}}{Label for the field}

\item{\code{type_field}}{Type of field (options : TODO)}

\item{\code{form_section}}{Left or right}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check_uniqueness_column_name"></a>}}
\if{latex}{\out{\hypertarget{method-check_uniqueness_column_name}{}}}
\subsection{Method \code{check_uniqueness_column_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$check_uniqueness_column_name(column)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-edit_label_field"></a>}}
\if{latex}{\out{\hypertarget{method-edit_label_field}{}}}
\subsection{Method \code{edit_label_field()}}{
Edit the label for an input field
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$edit_label_field(id_form, new_label)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-from_json"></a>}}
\if{latex}{\out{\hypertarget{method-from_json}{}}}
\subsection{Method \code{from_json()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$from_json(x, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_json"></a>}}
\if{latex}{\out{\hypertarget{method-to_json}{}}}
\subsection{Method \code{to_json()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$to_json(x, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-edit_options_field"></a>}}
\if{latex}{\out{\hypertarget{method-edit_options_field}{}}}
\subsection{Method \code{edit_options_field()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$edit_options_field(id_form, new_options)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_color"></a>}}
\if{latex}{\out{\hypertarget{method-is_color}{}}}
\subsection{Method \code{is_color()}}{
Is the provided color(s) valid?
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$is_color(colors)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-edit_options_colors"></a>}}
\if{latex}{\out{\hypertarget{method-edit_options_colors}{}}}
\subsection{Method \code{edit_options_colors()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$edit_options_colors(id_form, new_colors)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-amend_options_colors"></a>}}
\if{latex}{\out{\hypertarget{method-amend_options_colors}{}}}
\subsection{Method \code{amend_options_colors()}}{
If new options added, make sure the color vector is amended
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$amend_options_colors(id_form, options)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_options_order"></a>}}
\if{latex}{\out{\hypertarget{method-set_options_order}{}}}
\subsection{Method \code{set_options_order()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$set_options_order(id_form, new_order)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-amend_options_order"></a>}}
\if{latex}{\out{\hypertarget{method-amend_options_order}{}}}
\subsection{Method \code{amend_options_order()}}{
If new categories added, make sure the order vector is amended
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$amend_options_order(id_form, options)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-edit_formulier_setup"></a>}}
\if{latex}{\out{\hypertarget{method-edit_formulier_setup}{}}}
\subsection{Method \code{edit_formulier_setup()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$edit_formulier_setup(new_setup)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-edit_zichtbaarheid_invoerveld"></a>}}
\if{latex}{\out{\hypertarget{method-edit_zichtbaarheid_invoerveld}{}}}
\subsection{Method \code{edit_zichtbaarheid_invoerveld()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$edit_zichtbaarheid_invoerveld(id_formfield, new_zichtbaarheid)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-edit_verwijder_datum"></a>}}
\if{latex}{\out{\hypertarget{method-edit_verwijder_datum}{}}}
\subsection{Method \code{edit_verwijder_datum()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$edit_verwijder_datum(id_formfield, new_date)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-amend_formfield_order"></a>}}
\if{latex}{\out{\hypertarget{method-amend_formfield_order}{}}}
\subsection{Method \code{amend_formfield_order()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$amend_formfield_order(formside, deleted_number)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reset_volgorde_invoerveld"></a>}}
\if{latex}{\out{\hypertarget{method-reset_volgorde_invoerveld}{}}}
\subsection{Method \code{reset_volgorde_invoerveld()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$reset_volgorde_invoerveld(id_formfield, new_volgorde_nummer)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-column_type_from_field_type"></a>}}
\if{latex}{\out{\hypertarget{method-column_type_from_field_type}{}}}
\subsection{Method \code{column_type_from_field_type()}}{
Database column type based on field input type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$column_type_from_field_type(type)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prepare_data_table"></a>}}
\if{latex}{\out{\hypertarget{method-prepare_data_table}{}}}
\subsection{Method \code{prepare_data_table()}}{
Make sure the necessary columns are present in the data_table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$prepare_data_table()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-write_new_registration"></a>}}
\if{latex}{\out{\hypertarget{method-write_new_registration}{}}}
\subsection{Method \code{write_new_registration()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$write_new_registration(data, user_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-edit_registration"></a>}}
\if{latex}{\out{\hypertarget{method-edit_registration}{}}}
\subsection{Method \code{edit_registration()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$edit_registration(old_data, new_data, user_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read_registrations"></a>}}
\if{latex}{\out{\hypertarget{method-read_registrations}{}}}
\subsection{Method \code{read_registrations()}}{
Read registrations, recode select values where needed.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$read_registrations(recode = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{recode}}{If TRUE (default), replaces codes with labels (for select fields)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_registration_by_id"></a>}}
\if{latex}{\out{\hypertarget{method-get_registration_by_id}{}}}
\subsection{Method \code{get_registration_by_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$get_registration_by_id(id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-filter_period"></a>}}
\if{latex}{\out{\hypertarget{method-filter_period}{}}}
\subsection{Method \code{filter_period()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$filter_period(
  data,
  date_start = Sys.Date() - 7,
  date_end = Sys.Date()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read_definition"></a>}}
\if{latex}{\out{\hypertarget{method-read_definition}{}}}
\subsection{Method \code{read_definition()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$read_definition(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
