% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulierModule.R
\name{formulierUI}
\alias{formulierUI}
\alias{formulierModule}
\title{Formulier Shiny Module}
\usage{
formulierUI(
  id,
  title = "Formulier",
  icon = softui::bsicon("ui_checks"),
  header_ui = NULL,
  tag = "Opties",
  new_existing_labels = c("Nieuwe registratie invoeren",
    "Bestaande registratie bekijken")
)

formulierModule(input, output, session, .reg = NULL)
}
\description{
Formulier Shiny Module
}
