% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonFormSetupModule.R
\name{jsonFormSetupUI}
\alias{jsonFormSetupUI}
\alias{jsonFormSetupModule}
\title{Shiny module to adjust order of a form}
\usage{
jsonFormSetupUI(id, icon = NULL, label = "Edit values", status = "secondary")

jsonFormSetupModule(
  input,
  output,
  session,
  data = reactive(NULL),
  .reg = NULL,
  side_column = reactive(NULL),
  order_column = reactive(NULL),
  id_column = reactive(NULL),
  label_column = reactive(NULL),
  callback = function(data) { }
)
}
\description{
Shiny module to adjust order of a form
}
