% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_default_value.R
\name{make_default_value}
\alias{make_default_value}
\title{Make a default value for an input, based on data and column name}
\usage{
make_default_value(x, data, default = character(0), array = FALSE)
}
\description{
Make a default value for an input, based on data and column name
}
