% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registrationClass.R
\name{registrationClass}
\alias{registrationClass}
\title{R6 Class voor Registratie formulier}
\description{
R6 Class voor Registratie formulier

R6 Class voor Registratie formulier
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{registrationClass$new()}}
\item \href{#method-connect_to_database}{\code{registrationClass$connect_to_database()}}
\item \href{#method-close}{\code{registrationClass$close()}}
\item \href{#method-make_choices}{\code{registrationClass$make_choices()}}
\item \href{#method-choices_from_json}{\code{registrationClass$choices_from_json()}}
\item \href{#method-get_input_fields}{\code{registrationClass$get_input_fields()}}
\item \href{#method-get_next_formorder_number}{\code{registrationClass$get_next_formorder_number()}}
\item \href{#method-add_input_field_to_form}{\code{registrationClass$add_input_field_to_form()}}
\item \href{#method-check_uniqueness_column_name}{\code{registrationClass$check_uniqueness_column_name()}}
\item \href{#method-edit_label_field}{\code{registrationClass$edit_label_field()}}
\item \href{#method-from_json}{\code{registrationClass$from_json()}}
\item \href{#method-to_json}{\code{registrationClass$to_json()}}
\item \href{#method-edit_opties_invulveld}{\code{registrationClass$edit_opties_invulveld()}}
\item \href{#method-is_color}{\code{registrationClass$is_color()}}
\item \href{#method-set_optie_colors}{\code{registrationClass$set_optie_colors()}}
\item \href{#method-amend_optie_colors}{\code{registrationClass$amend_optie_colors()}}
\item \href{#method-set_optie_order}{\code{registrationClass$set_optie_order()}}
\item \href{#method-amend_optie_order}{\code{registrationClass$amend_optie_order()}}
\item \href{#method-edit_formulier_setup}{\code{registrationClass$edit_formulier_setup()}}
\item \href{#method-edit_zichtbaarheid_invoerveld}{\code{registrationClass$edit_zichtbaarheid_invoerveld()}}
\item \href{#method-edit_verwijder_datum}{\code{registrationClass$edit_verwijder_datum()}}
\item \href{#method-amend_formfield_order}{\code{registrationClass$amend_formfield_order()}}
\item \href{#method-reset_volgorde_invoerveld}{\code{registrationClass$reset_volgorde_invoerveld()}}
\item \href{#method-get_velden_form}{\code{registrationClass$get_velden_form()}}
\item \href{#method-clone}{\code{registrationClass$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Make new form object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{registrationClass$new(
  config_file = "conf/config.yml",
  what,
  schema = NULL,
  table = "formulier_velden",
  pool = TRUE,
  sqlite = NULL,
  default_color = "#3333cc"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config_file}}{Path to DB config}

\item{\code{schema}}{DB schema}

\item{\code{table}}{Table in DB (in schema) that holds form config}

\item{\code{pool}}{If TRUE, connects with dbPool}

\item{\code{sqlite}}{If path to an SQLite file, uses SQLite.}

\item{\code{default_color}}{Default color}

\item{\code{wat}}{Entry in config for DB connection}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connect_to_database"></a>}}
\if{latex}{\out{\hypertarget{method-connect_to_database}{}}}
\subsection{Method \code{connect_to_database()}}{
Connect to a database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{registrationClass$connect_to_database(
  config_file = NULL,
  schema = NULL,
  what = NULL,
  pool = TRUE,
  sqlite = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-close"></a>}}
\if{latex}{\out{\hypertarget{method-close}{}}}
\subsection{Method \code{close()}}{
Close database connection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{registrationClass$close()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-make_choices"></a>}}
\if{latex}{\out{\hypertarget{method-make_choices}{}}}
\subsection{Method \code{make_choices()}}{
Make choices (for selectInput) based on values and names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{registrationClass$make_choices(
  values_from,
  names_from = values_from,
  data = NULL,
  sort = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-choices_from_json"></a>}}
\if{latex}{\out{\hypertarget{method-choices_from_json}{}}}
\subsection{Method \code{choices_from_json()}}{
Unpack a JSON field to make a named vector
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{registrationClass$choices_from_json(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_input_fields"></a>}}
\if{latex}{\out{\hypertarget{method-get_input_fields}{}}}
\subsection{Method \code{get_input_fields()}}{
Get form input fields
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{registrationClass$get_input_fields(zichtbaarheid = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_next_formorder_number"></a>}}
\if{latex}{\out{\hypertarget{method-get_next_formorder_number}{}}}
\subsection{Method \code{get_next_formorder_number()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{registrationClass$get_next_formorder_number(
  kant_formulier = c("links", "rechts")
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_input_field_to_form"></a>}}
\if{latex}{\out{\hypertarget{method-add_input_field_to_form}{}}}
\subsection{Method \code{add_input_field_to_form()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{registrationClass$add_input_field_to_form(
  label_veld,
  type_veld,
  formulier_kant
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check_uniqueness_column_name"></a>}}
\if{latex}{\out{\hypertarget{method-check_uniqueness_column_name}{}}}
\subsection{Method \code{check_uniqueness_column_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{registrationClass$check_uniqueness_column_name(kolomnaam)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-edit_label_field"></a>}}
\if{latex}{\out{\hypertarget{method-edit_label_field}{}}}
\subsection{Method \code{edit_label_field()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{registrationClass$edit_label_field(id_formveld, new_label)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-from_json"></a>}}
\if{latex}{\out{\hypertarget{method-from_json}{}}}
\subsection{Method \code{from_json()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{registrationClass$from_json(x, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_json"></a>}}
\if{latex}{\out{\hypertarget{method-to_json}{}}}
\subsection{Method \code{to_json()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{registrationClass$to_json(x, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-edit_opties_invulveld"></a>}}
\if{latex}{\out{\hypertarget{method-edit_opties_invulveld}{}}}
\subsection{Method \code{edit_opties_invulveld()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{registrationClass$edit_opties_invulveld(id_formfield, new_opties)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_color"></a>}}
\if{latex}{\out{\hypertarget{method-is_color}{}}}
\subsection{Method \code{is_color()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{registrationClass$is_color(colors)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_optie_colors"></a>}}
\if{latex}{\out{\hypertarget{method-set_optie_colors}{}}}
\subsection{Method \code{set_optie_colors()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{registrationClass$set_optie_colors(id_formfield, new_colors)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-amend_optie_colors"></a>}}
\if{latex}{\out{\hypertarget{method-amend_optie_colors}{}}}
\subsection{Method \code{amend_optie_colors()}}{
If new options added, make sure the color vector is amended
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{registrationClass$amend_optie_colors(id_formfield, opties)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_optie_order"></a>}}
\if{latex}{\out{\hypertarget{method-set_optie_order}{}}}
\subsection{Method \code{set_optie_order()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{registrationClass$set_optie_order(id_formfield, new_order)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-amend_optie_order"></a>}}
\if{latex}{\out{\hypertarget{method-amend_optie_order}{}}}
\subsection{Method \code{amend_optie_order()}}{
If new categories added, make sure the order vector is amended
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{registrationClass$amend_optie_order(id_formfield, opties)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-edit_formulier_setup"></a>}}
\if{latex}{\out{\hypertarget{method-edit_formulier_setup}{}}}
\subsection{Method \code{edit_formulier_setup()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{registrationClass$edit_formulier_setup(new_setup)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-edit_zichtbaarheid_invoerveld"></a>}}
\if{latex}{\out{\hypertarget{method-edit_zichtbaarheid_invoerveld}{}}}
\subsection{Method \code{edit_zichtbaarheid_invoerveld()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{registrationClass$edit_zichtbaarheid_invoerveld(
  id_formfield,
  new_zichtbaarheid
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-edit_verwijder_datum"></a>}}
\if{latex}{\out{\hypertarget{method-edit_verwijder_datum}{}}}
\subsection{Method \code{edit_verwijder_datum()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{registrationClass$edit_verwijder_datum(id_formfield, new_date)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-amend_formfield_order"></a>}}
\if{latex}{\out{\hypertarget{method-amend_formfield_order}{}}}
\subsection{Method \code{amend_formfield_order()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{registrationClass$amend_formfield_order(formside, deleted_number)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reset_volgorde_invoerveld"></a>}}
\if{latex}{\out{\hypertarget{method-reset_volgorde_invoerveld}{}}}
\subsection{Method \code{reset_volgorde_invoerveld()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{registrationClass$reset_volgorde_invoerveld(id_formfield, new_volgorde_nummer)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_velden_form"></a>}}
\if{latex}{\out{\hypertarget{method-get_velden_form}{}}}
\subsection{Method \code{get_velden_form()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{registrationClass$get_velden_form(kant_formulier = c("links", "rechts"))}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{registrationClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
