% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registrationProcessMiningModule.R
\name{registrationProcessMiningUI}
\alias{registrationProcessMiningUI}
\alias{registrationProcessMiningModule}
\title{Shiny module for process mining visualization and analysis of status development of registration}
\usage{
registrationProcessMiningUI(id)

registrationProcessMiningModule(
  input,
  output,
  session,
  .reg,
  registration_id = reactive(NULL)
)
}
\arguments{
\item{id}{Shiny input id}

\item{.reg}{formClass for which the process mining is being done (call to formClass does not have event_data set as NULL)}

\item{registration_id}{registration about we want to see event traces and process mining analyses}
}
\description{
Shiny module for process mining visualization and analysis of status development of registration
}
