% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adminModule.R
\name{adminUI}
\alias{adminUI}
\alias{adminModule}
\title{Shiny module for form configuration}
\usage{
adminUI(id)

adminModule(input, output, session, .reg = NULL)
}
\description{
Shiny module for form configuration
}
