% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disable_inputs.R
\name{disable_inputs}
\alias{disable_inputs}
\title{Disable all inputs in a container}
\usage{
disable_inputs(id, session = shiny::getDefaultReactiveDomain())
}
\description{
Disable all inputs in a container
}
