% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formAdminModule.R
\name{formAdminUI}
\alias{formAdminUI}
\alias{formAdminModule}
\title{Shiny module for form configuration}
\usage{
formAdminUI(
  id,
  title = "Formulieropstelling",
  title_deleted = "Verwijderde invoervelden",
  header_ui = NULL
)

formAdminModule(input, output, session, .reg = NULL)
}
\arguments{
\item{id}{Shiny input id}

\item{title}{Title for the tab panel with settings}

\item{title_deleted}{Title for the tab panel with deleted fields}

\item{header_ui}{UI to add at the top of the panel with settings}
}
\description{
Shiny module for form configuration
}
