% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formClass.R
\name{formClass}
\alias{formClass}
\title{R6 Class voor Registratie formulier}
\description{
R6 Class voor Registratie formulier

R6 Class voor Registratie formulier
}
\section{Super class}{
\code{shintodb::NA}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NA-new}{\code{formClass$new()}}
\item \href{#method-NA-log}{\code{formClass$log()}}
\item \href{#method-NA-store_schema_str}{\code{formClass$store_schema_str()}}
\item \href{#method-NA-replace_value_where_multi}{\code{formClass$replace_value_where_multi()}}
\item \href{#method-NA-get_by_id}{\code{formClass$get_by_id()}}
\item \href{#method-NA-column_name_from_id}{\code{formClass$column_name_from_id()}}
\item \href{#method-NA-get_fields_by_type}{\code{formClass$get_fields_by_type()}}
\item \href{#method-NA-get_field_choices}{\code{formClass$get_field_choices()}}
\item \href{#method-NA-form_has_field}{\code{formClass$form_has_field()}}
\item \href{#method-NA-get_field_choices_in_order}{\code{formClass$get_field_choices_in_order()}}
\item \href{#method-NA-get_label_field}{\code{formClass$get_label_field()}}
\item \href{#method-NA-make_choices}{\code{formClass$make_choices()}}
\item \href{#method-NA-choices_from_json}{\code{formClass$choices_from_json()}}
\item \href{#method-NA-distinct_registration_field}{\code{formClass$distinct_registration_field()}}
\item \href{#method-NA-rename_definition_table}{\code{formClass$rename_definition_table()}}
\item \href{#method-NA-filter_by_visibility}{\code{formClass$filter_by_visibility()}}
\item \href{#method-NA-get_input_fields}{\code{formClass$get_input_fields()}}
\item \href{#method-NA-get_form_fields}{\code{formClass$get_form_fields()}}
\item \href{#method-NA-get_next_formorder_number}{\code{formClass$get_next_formorder_number()}}
\item \href{#method-NA-add_input_field_to_form}{\code{formClass$add_input_field_to_form()}}
\item \href{#method-NA-check_uniqueness_column_name}{\code{formClass$check_uniqueness_column_name()}}
\item \href{#method-NA-edit_label_field}{\code{formClass$edit_label_field()}}
\item \href{#method-NA-edit_field_type}{\code{formClass$edit_field_type()}}
\item \href{#method-NA-edit_filterable_column}{\code{formClass$edit_filterable_column()}}
\item \href{#method-NA-from_json}{\code{formClass$from_json()}}
\item \href{#method-NA-to_json}{\code{formClass$to_json()}}
\item \href{#method-NA-edit_options_field}{\code{formClass$edit_options_field()}}
\item \href{#method-NA-edit_nested_column_label}{\code{formClass$edit_nested_column_label()}}
\item \href{#method-NA-is_color}{\code{formClass$is_color()}}
\item \href{#method-NA-edit_options_colors}{\code{formClass$edit_options_colors()}}
\item \href{#method-NA-amend_options_colors}{\code{formClass$amend_options_colors()}}
\item \href{#method-NA-set_options_order}{\code{formClass$set_options_order()}}
\item \href{#method-NA-prepare_nested_choice_column}{\code{formClass$prepare_nested_choice_column()}}
\item \href{#method-NA-amend_options_order}{\code{formClass$amend_options_order()}}
\item \href{#method-NA-fix_options_order}{\code{formClass$fix_options_order()}}
\item \href{#method-NA-edit_formulier_setup}{\code{formClass$edit_formulier_setup()}}
\item \href{#method-NA-edit_zichtbaarheid_invoerveld}{\code{formClass$edit_zichtbaarheid_invoerveld()}}
\item \href{#method-NA-edit_verwijder_datum}{\code{formClass$edit_verwijder_datum()}}
\item \href{#method-NA-amend_formfield_order}{\code{formClass$amend_formfield_order()}}
\item \href{#method-NA-reset_volgorde_invoerveld}{\code{formClass$reset_volgorde_invoerveld()}}
\item \href{#method-NA-really_delete_formfield}{\code{formClass$really_delete_formfield()}}
\item \href{#method-NA-column_type_from_field_type}{\code{formClass$column_type_from_field_type()}}
\item \href{#method-NA-prepare_data_table}{\code{formClass$prepare_data_table()}}
\item \href{#method-NA-rename_registrations_intern}{\code{formClass$rename_registrations_intern()}}
\item \href{#method-NA-write_new_registration}{\code{formClass$write_new_registration()}}
\item \href{#method-NA-edit_registration}{\code{formClass$edit_registration()}}
\item \href{#method-NA-delete_registration}{\code{formClass$delete_registration()}}
\item \href{#method-NA-read_registrations}{\code{formClass$read_registrations()}}
\item \href{#method-NA-label_registrations_columns}{\code{formClass$label_registrations_columns()}}
\item \href{#method-NA-recode_registrations}{\code{formClass$recode_registrations()}}
\item \href{#method-NA-get_registration_by_id}{\code{formClass$get_registration_by_id()}}
\item \href{#method-NA-filter_period}{\code{formClass$filter_period()}}
\item \href{#method-NA-read_definition}{\code{formClass$read_definition()}}
\item \href{#method-NA-get_occurences}{\code{formClass$get_occurences()}}
\item \href{#method-NA-get_registration_by_json}{\code{formClass$get_registration_by_json()}}
\item \href{#method-NA-create_timeseries}{\code{formClass$create_timeseries()}}
\item \href{#method-NA-create_events}{\code{formClass$create_events()}}
\item \href{#method-NA-get_all_relations}{\code{formClass$get_all_relations()}}
\item \href{#method-NA-get_objects_for_collector}{\code{formClass$get_objects_for_collector()}}
\item \href{#method-NA-get_collector_for_object}{\code{formClass$get_collector_for_object()}}
\item \href{#method-NA-add_relation}{\code{formClass$add_relation()}}
\item \href{#method-NA-postgres_now}{\code{formClass$postgres_now()}}
\item \href{#method-NA-get_rows_since_auditstamp}{\code{formClass$get_rows_since_auditstamp()}}
\item \href{#method-NA-soft_delete_relations}{\code{formClass$soft_delete_relations()}}
\item \href{#method-NA-write_new_relations}{\code{formClass$write_new_relations()}}
\item \href{#method-NA-update_relations}{\code{formClass$update_relations()}}
\item \href{#method-unknown-clone}{\code{formClass$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="append_data"><a href='../../shintodb/html/NA.html#method-NA-append_data'><code>shintodb::NA$append_data()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="close"><a href='../../shintodb/html/NA.html#method-NA-close'><code>shintodb::NA$close()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="connect_to_database"><a href='../../shintodb/html/NA.html#method-NA-connect_to_database'><code>shintodb::NA$connect_to_database()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="delete_rows_where"><a href='../../shintodb/html/NA.html#method-NA-delete_rows_where'><code>shintodb::NA$delete_rows_where()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="execute_query"><a href='../../shintodb/html/NA.html#method-NA-execute_query'><code>shintodb::NA$execute_query()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="filter"><a href='../../shintodb/html/NA.html#method-NA-filter'><code>shintodb::NA$filter()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="get_query"><a href='../../shintodb/html/NA.html#method-NA-get_query'><code>shintodb::NA$get_query()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="has_connection"><a href='../../shintodb/html/NA.html#method-NA-has_connection'><code>shintodb::NA$has_connection()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="has_value"><a href='../../shintodb/html/NA.html#method-NA-has_value'><code>shintodb::NA$has_value()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="have_column"><a href='../../shintodb/html/NA.html#method-NA-have_column'><code>shintodb::NA$have_column()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="list_indexes"><a href='../../shintodb/html/NA.html#method-NA-list_indexes'><code>shintodb::NA$list_indexes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="list_tables"><a href='../../shintodb/html/NA.html#method-NA-list_tables'><code>shintodb::NA$list_tables()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="make_column"><a href='../../shintodb/html/NA.html#method-NA-make_column'><code>shintodb::NA$make_column()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="postgres_now_in_local_time"><a href='../../shintodb/html/NA.html#method-NA-postgres_now_in_local_time'><code>shintodb::NA$postgres_now_in_local_time()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="query"><a href='../../shintodb/html/NA.html#method-NA-query'><code>shintodb::NA$query()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="read_table"><a href='../../shintodb/html/NA.html#method-NA-read_table'><code>shintodb::NA$read_table()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="replace_value_where"><a href='../../shintodb/html/NA.html#method-NA-replace_value_where'><code>shintodb::NA$replace_value_where()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="replace_value_where2"><a href='../../shintodb/html/NA.html#method-NA-replace_value_where2'><code>shintodb::NA$replace_value_where2()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="table_columns"><a href='../../shintodb/html/NA.html#method-NA-table_columns'><code>shintodb::NA$table_columns()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="table_info"><a href='../../shintodb/html/NA.html#method-NA-table_info'><code>shintodb::NA$table_info()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="update_where"><a href='../../shintodb/html/NA.html#method-NA-update_where'><code>shintodb::NA$update_where()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="write_table_overwrite"><a href='../../shintodb/html/NA.html#method-NA-write_table_overwrite'><code>shintodb::NA$write_table_overwrite()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-new"></a>}}
\if{latex}{\out{\hypertarget{method-NA-new}{}}}
\subsection{Method \code{new()}}{
Make new form object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$new(
  config_file = "conf/config.yml",
  what,
  schema = NULL,
  class_type = "",
  db_connection = NULL,
  filterable = FALSE,
  audit = FALSE,
  audit_table = "registrations_audit",
  def_table = "formulier_velden",
  def_columns = list(id_form = "id_formulierveld", column_field = "kolomnaam_veld",
    label_field = "label_veld", type_field = "type_veld", order_field = "volgorde_veld",
    form_section = "formulier_sectie", visible = "zichtbaar", date_deleted =
    "datum_uitgeschakeld", options = "opties", order_options = "volgorde_opties", colors
    = "kleuren", removable = "kan_worden_verwijderd", make_filter = "make_filter",
    tooltip = "tooltip"),
  data_table = NULL,
  data_columns = list(id = "id_registratie", name = "naam_registratie", time_created =
    "invoerdatum", time_modified = "wijzigdatum", user = "user_id", status = "status",
    priority = "priority", deleted = "deleted"),
  relation_table = NULL,
  relation_columns = list(id = "id", collector_id = "collector_id", collector_type =
    "collector_type", object_id = "object_id", object_type = "object_type", relation_id =
    "relation_id", relation_type = "relation_type", comment = "comment", status =
    "status", username = "username", timestamp = "timestamp", verwijderd = "verwijderd"),
  relation_audit_table = NULL,
  inject = NULL,
  pool = TRUE,
  sqlite = NULL,
  connect_on_init = TRUE,
  shintousers_object = NULL,
  default_color = "#3333cc"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config_file}}{Path to DB config}

\item{\code{what}}{Entry in config for DB connection}

\item{\code{schema}}{DB schema}

\item{\code{class_type}}{Name for the class of the objects to be managed with shintoforms}

\item{\code{db_connection}}{Optional, an existing DB connection}

\item{\code{filterable}}{Boolean, TRUE if the user wants to control which registrations to filter. Requires make_filter and tooltip in the def_columns list.}

\item{\code{audit}}{If TRUE, writes edits to audit log}

\item{\code{audit_table}}{Name of the table for the audit logs}

\item{\code{def_table}}{Table in DB (in schema) that holds form config}

\item{\code{def_columns}}{List with internal names of the config fields, and names of the corresponding columns in the table}

\item{\code{data_table}}{Name of the table to hold the data saved by shintoforms}

\item{\code{data_columns}}{List with internal names of required data columns, and corresponding names in the data table}

\item{\code{relation_table}}{Name of the 'relations' table}

\item{\code{relation_columns}}{List with internal names of the relations table and corresponding names in the actual postgres table}

\item{\code{relation_audit_table}}{If TRUE, audits the relation table}

\item{\code{inject}}{Custom fields to be injected; see Apollo for examples}

\item{\code{pool}}{If TRUE, connects with dbPool}

\item{\code{sqlite}}{If path to an SQLite file, uses SQLite.}

\item{\code{connect_on_init}}{If TRUE (default), connects to DB on initialization. Otherwise, call \verb{$connect_to_database} at a later stage.}

\item{\code{shintousers_object}}{Optional, an instance of shintoUser (".user") passed during init.}

\item{\code{default_color}}{Default color}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-log"></a>}}
\if{latex}{\out{\hypertarget{method-NA-log}{}}}
\subsection{Method \code{log()}}{
Log something
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$log(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-store_schema_str"></a>}}
\if{latex}{\out{\hypertarget{method-NA-store_schema_str}{}}}
\subsection{Method \code{store_schema_str()}}{
(Re-)write schema_str
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$store_schema_str()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-replace_value_where_multi"></a>}}
\if{latex}{\out{\hypertarget{method-NA-replace_value_where_multi}{}}}
\subsection{Method \code{replace_value_where_multi()}}{
Multi-replace value where (sql update wrapper)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$replace_value_where_multi(
  table,
  replace_list,
  col_compare,
  val_compare,
  query_only = FALSE,
  quiet = FALSE,
  username = ""
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{table}}{Table name (no schema)}

\item{\code{replace_list}}{Named list of values to replace; name of list element is column name}

\item{\code{col_compare}}{Which column to compare (e.g. 'id')}

\item{\code{val_compare}}{Value to compare (e.g. 101, then id = 101 will be updated according to replace_list)}

\item{\code{query_only}}{If TRUE, returns only the qyuery}

\item{\code{quiet}}{If TRUE, prints nothing}

\item{\code{username}}{If passed; updates 'username' in table for audit purposes}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-get_by_id"></a>}}
\if{latex}{\out{\hypertarget{method-NA-get_by_id}{}}}
\subsection{Method \code{get_by_id()}}{
Get a row from the form definition by the form id.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$get_by_id(id_form)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id_form}}{Id of form field}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-column_name_from_id"></a>}}
\if{latex}{\out{\hypertarget{method-NA-column_name_from_id}{}}}
\subsection{Method \code{column_name_from_id()}}{
Get column name for a form field by ID
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$column_name_from_id(id_form)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id_form}}{Id of form field}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-get_fields_by_type"></a>}}
\if{latex}{\out{\hypertarget{method-NA-get_fields_by_type}{}}}
\subsection{Method \code{get_fields_by_type()}}{
Get all fields of a certain type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$get_fields_by_type(field_type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{field_type}}{Type of field, e.g. 'singleselect'}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-get_field_choices"></a>}}
\if{latex}{\out{\hypertarget{method-NA-get_field_choices}{}}}
\subsection{Method \code{get_field_choices()}}{
Get (recoded) choices for a select field
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$get_field_choices(column_field)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{column_field}}{Name of field (i.e. column name)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-form_has_field"></a>}}
\if{latex}{\out{\hypertarget{method-NA-form_has_field}{}}}
\subsection{Method \code{form_has_field()}}{
Check if we have a field with a name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$form_has_field(column_field)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-get_field_choices_in_order"></a>}}
\if{latex}{\out{\hypertarget{method-NA-get_field_choices_in_order}{}}}
\subsection{Method \code{get_field_choices_in_order()}}{
Get order of the choices for a field
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$get_field_choices_in_order(column_field)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{column_field}}{Name of field (i.e. column name)}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Safe implementation; if the order length does not match the options,
or it is otherwise corrupted, a simple vector of length(n choices) is returned.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-get_label_field"></a>}}
\if{latex}{\out{\hypertarget{method-NA-get_label_field}{}}}
\subsection{Method \code{get_label_field()}}{
Get the label for a field
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$get_label_field(column_field)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{column_field}}{Name of field (i.e. column name)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-make_choices"></a>}}
\if{latex}{\out{\hypertarget{method-NA-make_choices}{}}}
\subsection{Method \code{make_choices()}}{
Make choices (for selectInput) based on values and names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$make_choices(
  values_from,
  names_from = values_from,
  data = NULL,
  sort = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{values_from}}{Name of column to take values from}

\item{\code{names_from}}{Name of column to take names from}

\item{\code{data}}{Dataframe}

\item{\code{sort}}{If TRUE, sorts choices alphabetically}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-choices_from_json"></a>}}
\if{latex}{\out{\hypertarget{method-NA-choices_from_json}{}}}
\subsection{Method \code{choices_from_json()}}{
Unpack a JSON field to make a named vector
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$choices_from_json(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A JSON string}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-distinct_registration_field"></a>}}
\if{latex}{\out{\hypertarget{method-NA-distinct_registration_field}{}}}
\subsection{Method \code{distinct_registration_field()}}{
Get distinct values of a column in the registrations table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$distinct_registration_field(column_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{column_name}}{Name of column in data_table}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-rename_definition_table"></a>}}
\if{latex}{\out{\hypertarget{method-NA-rename_definition_table}{}}}
\subsection{Method \code{rename_definition_table()}}{
Rename database table to correct internal column names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$rename_definition_table(data, from_where = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Dataframe}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-filter_by_visibility"></a>}}
\if{latex}{\out{\hypertarget{method-NA-filter_by_visibility}{}}}
\subsection{Method \code{filter_by_visibility()}}{
Filter fields by visibility (or not)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$filter_by_visibility(data, visibility = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Dataframe}

\item{\code{visibility}}{TRUE or FALSE. When NULL, returns all data.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-get_input_fields"></a>}}
\if{latex}{\out{\hypertarget{method-NA-get_input_fields}{}}}
\subsection{Method \code{get_input_fields()}}{
Get all non-deleted form input fields
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$get_input_fields(zichtbaarheid = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{zichtbaarheid}}{Visibility}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-get_form_fields"></a>}}
\if{latex}{\out{\hypertarget{method-NA-get_form_fields}{}}}
\subsection{Method \code{get_form_fields()}}{
Get form fields for left or right column of the form
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$get_form_fields(form_section)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{form_section}}{Either 1 (left) or 2 (right)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-get_next_formorder_number"></a>}}
\if{latex}{\out{\hypertarget{method-NA-get_next_formorder_number}{}}}
\subsection{Method \code{get_next_formorder_number()}}{
Get next available form order number
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$get_next_formorder_number(form_section)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{form_section}}{Either 1 (left) or 2 (right)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-add_input_field_to_form"></a>}}
\if{latex}{\out{\hypertarget{method-NA-add_input_field_to_form}{}}}
\subsection{Method \code{add_input_field_to_form()}}{
Adds an input field to a form
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$add_input_field_to_form(
  label_field,
  type_field,
  form_section = NULL,
  filterable = NULL,
  tooltip = NULL,
  column_name = NULL,
  column_2_name = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{label_field}}{Label for the field}

\item{\code{type_field}}{Type of field (options : TODO)}

\item{\code{form_section}}{Either 1 (left) or 2 (right)}

\item{\code{filterable}}{Whether it is a filterable field (not used)}

\item{\code{tooltip}}{For filter, the tooltip text (not used)}

\item{\code{column_name}}{Optional; name of column in output data (otherwise made from label_field)}

\item{\code{column_2_name}}{Only used for nested select}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-check_uniqueness_column_name"></a>}}
\if{latex}{\out{\hypertarget{method-NA-check_uniqueness_column_name}{}}}
\subsection{Method \code{check_uniqueness_column_name()}}{
Check if a column name can be used (returns TRUE if the column does not exist in the definition table)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$check_uniqueness_column_name(column)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{column}}{Name of the column}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-edit_label_field"></a>}}
\if{latex}{\out{\hypertarget{method-NA-edit_label_field}{}}}
\subsection{Method \code{edit_label_field()}}{
Edit the label for an input field
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$edit_label_field(id_form, new_label)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id_form}}{ID of form field}

\item{\code{new_label}}{New label text}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-edit_field_type"></a>}}
\if{latex}{\out{\hypertarget{method-NA-edit_field_type}{}}}
\subsection{Method \code{edit_field_type()}}{
Change the field type for a form field
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$edit_field_type(id_form, new_type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id_form}}{ID of form field}

\item{\code{new_type}}{New type}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-edit_filterable_column"></a>}}
\if{latex}{\out{\hypertarget{method-NA-edit_filterable_column}{}}}
\subsection{Method \code{edit_filterable_column()}}{
Edit a filterable field. Not used anymore.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$edit_filterable_column(id_form, new_filter_boolean, new_tooltip)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id_form}}{ID of form field}

\item{\code{new_filter_boolean}}{New value}

\item{\code{new_tooltip}}{New tooltip}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-from_json"></a>}}
\if{latex}{\out{\hypertarget{method-NA-from_json}{}}}
\subsection{Method \code{from_json()}}{
from_json wrapper
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$from_json(x, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Text}

\item{\code{...}}{Other args to \code{\link[shintocatman:json]{shintocatman::from_json()}}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-to_json"></a>}}
\if{latex}{\out{\hypertarget{method-NA-to_json}{}}}
\subsection{Method \code{to_json()}}{
to_json wrapper
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$to_json(x, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Text}

\item{\code{...}}{Other args to \code{\link[shintocatman:json]{shintocatman::to_json()}}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-edit_options_field"></a>}}
\if{latex}{\out{\hypertarget{method-NA-edit_options_field}{}}}
\subsection{Method \code{edit_options_field()}}{
Edit options (choices) for a form field
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$edit_options_field(id_form, new_options)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-edit_nested_column_label"></a>}}
\if{latex}{\out{\hypertarget{method-NA-edit_nested_column_label}{}}}
\subsection{Method \code{edit_nested_column_label()}}{
Label for nested column is kept in options list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$edit_nested_column_label(id_form, opts, new_label)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id_form}}{ID of form field}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-is_color"></a>}}
\if{latex}{\out{\hypertarget{method-NA-is_color}{}}}
\subsection{Method \code{is_color()}}{
Is the provided color(s) valid?
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$is_color(colors)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-edit_options_colors"></a>}}
\if{latex}{\out{\hypertarget{method-NA-edit_options_colors}{}}}
\subsection{Method \code{edit_options_colors()}}{
Edit colors for a form field
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$edit_options_colors(id_form, new_colors)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id_form}}{ID of form field}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-amend_options_colors"></a>}}
\if{latex}{\out{\hypertarget{method-NA-amend_options_colors}{}}}
\subsection{Method \code{amend_options_colors()}}{
If new options added, make sure the color vector is amended
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$amend_options_colors(id_form, options)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id_form}}{ID of form field}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-set_options_order"></a>}}
\if{latex}{\out{\hypertarget{method-NA-set_options_order}{}}}
\subsection{Method \code{set_options_order()}}{
Set the order for a single/multi select field
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$set_options_order(id_form, new_order)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id_form}}{ID of form field}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-prepare_nested_choice_column"></a>}}
\if{latex}{\out{\hypertarget{method-NA-prepare_nested_choice_column}{}}}
\subsection{Method \code{prepare_nested_choice_column()}}{
Prepare a nested choice column
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$prepare_nested_choice_column(id_form, name, options)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id_form}}{ID of form field}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-amend_options_order"></a>}}
\if{latex}{\out{\hypertarget{method-NA-amend_options_order}{}}}
\subsection{Method \code{amend_options_order()}}{
If new categories added or removed, make sure the order vector is amended
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$amend_options_order(id_form, options)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id_form}}{ID of form field}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-fix_options_order"></a>}}
\if{latex}{\out{\hypertarget{method-NA-fix_options_order}{}}}
\subsection{Method \code{fix_options_order()}}{
Like amend_options_order but takes field name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$fix_options_order(column_field)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-edit_formulier_setup"></a>}}
\if{latex}{\out{\hypertarget{method-NA-edit_formulier_setup}{}}}
\subsection{Method \code{edit_formulier_setup()}}{
Edit the form layout
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$edit_formulier_setup(new_setup)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-edit_zichtbaarheid_invoerveld"></a>}}
\if{latex}{\out{\hypertarget{method-NA-edit_zichtbaarheid_invoerveld}{}}}
\subsection{Method \code{edit_zichtbaarheid_invoerveld()}}{
Change the visibility
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$edit_zichtbaarheid_invoerveld(id_formfield, new_zichtbaarheid)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id_formfield}}{ID of form field}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-edit_verwijder_datum"></a>}}
\if{latex}{\out{\hypertarget{method-NA-edit_verwijder_datum}{}}}
\subsection{Method \code{edit_verwijder_datum()}}{
Change the removal date
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$edit_verwijder_datum(id_formfield, new_date)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id_formfield}}{ID of form field}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-amend_formfield_order"></a>}}
\if{latex}{\out{\hypertarget{method-NA-amend_formfield_order}{}}}
\subsection{Method \code{amend_formfield_order()}}{
Change the formfield order
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$amend_formfield_order(formside, deleted_number)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-reset_volgorde_invoerveld"></a>}}
\if{latex}{\out{\hypertarget{method-NA-reset_volgorde_invoerveld}{}}}
\subsection{Method \code{reset_volgorde_invoerveld()}}{
Reset the form field order
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$reset_volgorde_invoerveld(id_formfield, new_volgorde_nummer)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id_formfield}}{ID of form field}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-really_delete_formfield"></a>}}
\if{latex}{\out{\hypertarget{method-NA-really_delete_formfield}{}}}
\subsection{Method \code{really_delete_formfield()}}{
Actually really delete a field
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$really_delete_formfield(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{ID of form field}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-column_type_from_field_type"></a>}}
\if{latex}{\out{\hypertarget{method-NA-column_type_from_field_type}{}}}
\subsection{Method \code{column_type_from_field_type()}}{
Database column type based on field input type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$column_type_from_field_type(type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{One of 'text', 'numeric', 'boolean', 'singleselect', 'multiselect', 'date', 'singlecheck'}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-prepare_data_table"></a>}}
\if{latex}{\out{\hypertarget{method-NA-prepare_data_table}{}}}
\subsection{Method \code{prepare_data_table()}}{
Make sure the necessary columns are present in the data_table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$prepare_data_table()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-rename_registrations_intern"></a>}}
\if{latex}{\out{\hypertarget{method-NA-rename_registrations_intern}{}}}
\subsection{Method \code{rename_registrations_intern()}}{
Rename registrations data (as read with $read_registrations) to
internal column names as defined in \code{data_columns} argument
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$rename_registrations_intern(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Dataframe}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-write_new_registration"></a>}}
\if{latex}{\out{\hypertarget{method-NA-write_new_registration}{}}}
\subsection{Method \code{write_new_registration()}}{
Writes a registration to the data table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$write_new_registration(
  data,
  user_id,
  current_reg_id,
  data_only = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{data}

\item{\code{user_id}}{User ID}

\item{\code{current_reg_id}}{ID for the registration}

\item{\code{data_only}}{Return only the dataset before appending to an actual table (for testing)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-edit_registration"></a>}}
\if{latex}{\out{\hypertarget{method-NA-edit_registration}{}}}
\subsection{Method \code{edit_registration()}}{
Edits an existing registration
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$edit_registration(old_data, new_data, user_id, current_reg_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{old_data}}{Dataframe with previous entry of registration}

\item{\code{new_data}}{Dataframe with new data}

\item{\code{user_id}}{User ID, for audit logging}

\item{\code{current_reg_id}}{Registration ID}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-delete_registration"></a>}}
\if{latex}{\out{\hypertarget{method-NA-delete_registration}{}}}
\subsection{Method \code{delete_registration()}}{
Soft/hard/(un)delete a registration. \code{deleted} field will be set to 1.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$delete_registration(
  registration_id,
  method = c("soft", "hard", "undelete")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{registration_id}}{ID of the registration}

\item{\code{method}}{Either 'soft', 'hard' (really delete), 'undelete'}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-read_registrations"></a>}}
\if{latex}{\out{\hypertarget{method-NA-read_registrations}{}}}
\subsection{Method \code{read_registrations()}}{
Read registrations, recode select values where needed.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$read_registrations(
  recode = TRUE,
  include_deleted = TRUE,
  deleted_only = FALSE,
  lazy = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{recode}}{If TRUE (default), replaces codes with labels (for select fields)}

\item{\code{include_deleted}}{If TRUE, include deleted registrations}

\item{\code{lazy}}{If TRUE, returns a lazy tbl}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-label_registrations_columns"></a>}}
\if{latex}{\out{\hypertarget{method-NA-label_registrations_columns}{}}}
\subsection{Method \code{label_registrations_columns()}}{
Replace column names of registrations with their labels
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$label_registrations_columns(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Read in with \code{read_registrations} or \code{get_registration_by_id}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-recode_registrations"></a>}}
\if{latex}{\out{\hypertarget{method-NA-recode_registrations}{}}}
\subsection{Method \code{recode_registrations()}}{
Recode select values in registrations data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$recode_registrations(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Dataframe (registrations data)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-get_registration_by_id"></a>}}
\if{latex}{\out{\hypertarget{method-NA-get_registration_by_id}{}}}
\subsection{Method \code{get_registration_by_id()}}{
Get a registration with a uuid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$get_registration_by_id(id, recode = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{ID of the registration (uuid)}

\item{\code{recode}}{If TRUE, recodes integer values to their labels}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-filter_period"></a>}}
\if{latex}{\out{\hypertarget{method-NA-filter_period}{}}}
\subsection{Method \code{filter_period()}}{
Filter data by time_created
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$filter_period(
  data,
  date_start = Sys.Date() - 7,
  date_end = Sys.Date()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Read with \code{read_registrations}}

\item{\code{date_start}}{Date start}

\item{\code{date_end}}{Date end}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-read_definition"></a>}}
\if{latex}{\out{\hypertarget{method-NA-read_definition}{}}}
\subsection{Method \code{read_definition()}}{
Reads the definition table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$read_definition(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-get_occurences"></a>}}
\if{latex}{\out{\hypertarget{method-NA-get_occurences}{}}}
\subsection{Method \code{get_occurences()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$get_occurences(table, column, record)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-get_registration_by_json"></a>}}
\if{latex}{\out{\hypertarget{method-NA-get_registration_by_json}{}}}
\subsection{Method \code{get_registration_by_json()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$get_registration_by_json(table, column, record)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-create_timeseries"></a>}}
\if{latex}{\out{\hypertarget{method-NA-create_timeseries}{}}}
\subsection{Method \code{create_timeseries()}}{
Creates a timeseries for the audit
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$create_timeseries(columns = NULL, date_range = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-create_events"></a>}}
\if{latex}{\out{\hypertarget{method-NA-create_events}{}}}
\subsection{Method \code{create_events()}}{
Creates 'events' based on the timeseries
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$create_events(timeseries, date_range = NULL, creation_only = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-get_all_relations"></a>}}
\if{latex}{\out{\hypertarget{method-NA-get_all_relations}{}}}
\subsection{Method \code{get_all_relations()}}{
Reads the relation table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$get_all_relations()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-get_objects_for_collector"></a>}}
\if{latex}{\out{\hypertarget{method-NA-get_objects_for_collector}{}}}
\subsection{Method \code{get_objects_for_collector()}}{
get_objects_for_collector
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$get_objects_for_collector(
  collector_type = NA,
  collector_id = NA,
  relation_type = NA,
  relation_id = NA,
  object_type = NA,
  filter_verwijderd = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{collector_type}}{Type of collector}

\item{\code{collector_id}}{ID of collector}

\item{\code{relation_type}}{Description of collector -> object relation}

\item{\code{relation_id}}{ID of relation (often session ID of module)}

\item{\code{filter_verwijderd}}{when status is zero}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-get_collector_for_object"></a>}}
\if{latex}{\out{\hypertarget{method-NA-get_collector_for_object}{}}}
\subsection{Method \code{get_collector_for_object()}}{
get_collector_for_object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$get_collector_for_object(
  object_id,
  collector_type = NA,
  object_type = NA,
  relation_type = NA,
  relation_id = NA,
  filter_verwijderd = T
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{collector_type}}{Type of collector}

\item{\code{relation_type}}{Description of collector -> object relation}

\item{\code{relation_id}}{ID of relation (often session ID of module)}

\item{\code{filter_verwijderd}}{when status is zero}

\item{\code{collector_id}}{ID of collector}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-add_relation"></a>}}
\if{latex}{\out{\hypertarget{method-NA-add_relation}{}}}
\subsection{Method \code{add_relation()}}{
Adds a relation to the relation table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$add_relation(
  id = uuid::UUIDgenerate(),
  collector_id,
  collector_type,
  object_id,
  object_type,
  relation_id,
  relation_type,
  comment = NA,
  status = "1",
  verwijderd = FALSE,
  username,
  timestamp
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-postgres_now"></a>}}
\if{latex}{\out{\hypertarget{method-NA-postgres_now}{}}}
\subsection{Method \code{postgres_now()}}{
timestamp in postgres timezone
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$postgres_now()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-get_rows_since_auditstamp"></a>}}
\if{latex}{\out{\hypertarget{method-NA-get_rows_since_auditstamp}{}}}
\subsection{Method \code{get_rows_since_auditstamp()}}{
Select rows in registration audit table since some timestamp
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$get_rows_since_auditstamp(time_since)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{time_since}}{POSIXct timestamp}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-soft_delete_relations"></a>}}
\if{latex}{\out{\hypertarget{method-NA-soft_delete_relations}{}}}
\subsection{Method \code{soft_delete_relations()}}{
Soft delete relations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$soft_delete_relations(ids)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ids}}{ID of rows to delete}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-write_new_relations"></a>}}
\if{latex}{\out{\hypertarget{method-NA-write_new_relations}{}}}
\subsection{Method \code{write_new_relations()}}{
Append relations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$write_new_relations(data, registration_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Data to append}

\item{\code{registration_id}}{ID for the registration}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-update_relations"></a>}}
\if{latex}{\out{\hypertarget{method-NA-update_relations}{}}}
\subsection{Method \code{update_relations()}}{
update current relation table for registration ID
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$update_relations(new_relations, registration_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_relations}}{The current state that should end up in the relations table}

\item{\code{registration_id}}{ID for the registration}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-unknown-clone"></a>}}
\if{latex}{\out{\hypertarget{method-unknown-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
