% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formClass.R
\name{formClass}
\alias{formClass}
\title{R6 Class voor Registratie formulier}
\description{
R6 Class voor Registratie formulier

R6 Class voor Registratie formulier
}
\section{Super class}{
\code{shintodb::NA}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NA-new}{\code{formClass$new()}}
\item \href{#method-NA-log}{\code{formClass$log()}}
\item \href{#method-NA-replace_value_where_multi}{\code{formClass$replace_value_where_multi()}}
\item \href{#method-NA-get_by_id}{\code{formClass$get_by_id()}}
\item \href{#method-NA-column_name_from_id}{\code{formClass$column_name_from_id()}}
\item \href{#method-NA-get_fields_by_type}{\code{formClass$get_fields_by_type()}}
\item \href{#method-NA-get_field_choices}{\code{formClass$get_field_choices()}}
\item \href{#method-NA-make_choices}{\code{formClass$make_choices()}}
\item \href{#method-NA-choices_from_json}{\code{formClass$choices_from_json()}}
\item \href{#method-NA-distinct_registration_field}{\code{formClass$distinct_registration_field()}}
\item \href{#method-NA-rename_definition_table}{\code{formClass$rename_definition_table()}}
\item \href{#method-NA-get_input_fields}{\code{formClass$get_input_fields()}}
\item \href{#method-NA-get_form_fields}{\code{formClass$get_form_fields()}}
\item \href{#method-NA-get_next_formorder_number}{\code{formClass$get_next_formorder_number()}}
\item \href{#method-NA-add_input_field_to_form}{\code{formClass$add_input_field_to_form()}}
\item \href{#method-NA-check_uniqueness_column_name}{\code{formClass$check_uniqueness_column_name()}}
\item \href{#method-NA-edit_label_field}{\code{formClass$edit_label_field()}}
\item \href{#method-NA-edit_field_type}{\code{formClass$edit_field_type()}}
\item \href{#method-NA-edit_filterable_column}{\code{formClass$edit_filterable_column()}}
\item \href{#method-NA-from_json}{\code{formClass$from_json()}}
\item \href{#method-NA-to_json}{\code{formClass$to_json()}}
\item \href{#method-NA-edit_options_field}{\code{formClass$edit_options_field()}}
\item \href{#method-NA-edit_nested_column_label}{\code{formClass$edit_nested_column_label()}}
\item \href{#method-NA-is_color}{\code{formClass$is_color()}}
\item \href{#method-NA-edit_options_colors}{\code{formClass$edit_options_colors()}}
\item \href{#method-NA-amend_options_colors}{\code{formClass$amend_options_colors()}}
\item \href{#method-NA-set_options_order}{\code{formClass$set_options_order()}}
\item \href{#method-NA-prepare_nested_choice_column}{\code{formClass$prepare_nested_choice_column()}}
\item \href{#method-NA-amend_options_order}{\code{formClass$amend_options_order()}}
\item \href{#method-NA-edit_formulier_setup}{\code{formClass$edit_formulier_setup()}}
\item \href{#method-NA-edit_zichtbaarheid_invoerveld}{\code{formClass$edit_zichtbaarheid_invoerveld()}}
\item \href{#method-NA-edit_verwijder_datum}{\code{formClass$edit_verwijder_datum()}}
\item \href{#method-NA-amend_formfield_order}{\code{formClass$amend_formfield_order()}}
\item \href{#method-NA-reset_volgorde_invoerveld}{\code{formClass$reset_volgorde_invoerveld()}}
\item \href{#method-NA-really_delete_formfield}{\code{formClass$really_delete_formfield()}}
\item \href{#method-NA-column_type_from_field_type}{\code{formClass$column_type_from_field_type()}}
\item \href{#method-NA-prepare_data_table}{\code{formClass$prepare_data_table()}}
\item \href{#method-NA-rename_registrations_intern}{\code{formClass$rename_registrations_intern()}}
\item \href{#method-NA-write_new_registration}{\code{formClass$write_new_registration()}}
\item \href{#method-NA-edit_registration}{\code{formClass$edit_registration()}}
\item \href{#method-NA-delete_registration}{\code{formClass$delete_registration()}}
\item \href{#method-NA-read_registrations}{\code{formClass$read_registrations()}}
\item \href{#method-NA-label_registrations_columns}{\code{formClass$label_registrations_columns()}}
\item \href{#method-NA-recode_registrations}{\code{formClass$recode_registrations()}}
\item \href{#method-NA-get_registration_by_id}{\code{formClass$get_registration_by_id()}}
\item \href{#method-NA-filter_period}{\code{formClass$filter_period()}}
\item \href{#method-NA-read_definition}{\code{formClass$read_definition()}}
\item \href{#method-NA-get_occurences}{\code{formClass$get_occurences()}}
\item \href{#method-NA-get_registration_by_json}{\code{formClass$get_registration_by_json()}}
\item \href{#method-NA-create_timeseries}{\code{formClass$create_timeseries()}}
\item \href{#method-NA-create_events}{\code{formClass$create_events()}}
\item \href{#method-NA-get_all_relations}{\code{formClass$get_all_relations()}}
\item \href{#method-NA-get_objects_for_collector}{\code{formClass$get_objects_for_collector()}}
\item \href{#method-NA-get_collector_for_object}{\code{formClass$get_collector_for_object()}}
\item \href{#method-NA-add_relation}{\code{formClass$add_relation()}}
\item \href{#method-NA-postgres_now}{\code{formClass$postgres_now()}}
\item \href{#method-NA-get_rows_since_auditstamp}{\code{formClass$get_rows_since_auditstamp()}}
\item \href{#method-NA-soft_delete_relations}{\code{formClass$soft_delete_relations()}}
\item \href{#method-NA-write_new_relations}{\code{formClass$write_new_relations()}}
\item \href{#method-NA-update_relations}{\code{formClass$update_relations()}}
\item \href{#method-unknown-clone}{\code{formClass$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="append_data"><a href='../../shintodb/html/NA.html#method-NA-append_data'><code>shintodb::NA$append_data()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="close"><a href='../../shintodb/html/NA.html#method-NA-close'><code>shintodb::NA$close()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="connect_to_database"><a href='../../shintodb/html/NA.html#method-NA-connect_to_database'><code>shintodb::NA$connect_to_database()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="delete_rows_where"><a href='../../shintodb/html/NA.html#method-NA-delete_rows_where'><code>shintodb::NA$delete_rows_where()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="execute_query"><a href='../../shintodb/html/NA.html#method-NA-execute_query'><code>shintodb::NA$execute_query()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="filter"><a href='../../shintodb/html/NA.html#method-NA-filter'><code>shintodb::NA$filter()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="get_query"><a href='../../shintodb/html/NA.html#method-NA-get_query'><code>shintodb::NA$get_query()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="has_connection"><a href='../../shintodb/html/NA.html#method-NA-has_connection'><code>shintodb::NA$has_connection()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="has_value"><a href='../../shintodb/html/NA.html#method-NA-has_value'><code>shintodb::NA$has_value()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="have_column"><a href='../../shintodb/html/NA.html#method-NA-have_column'><code>shintodb::NA$have_column()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="list_indexes"><a href='../../shintodb/html/NA.html#method-NA-list_indexes'><code>shintodb::NA$list_indexes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="list_tables"><a href='../../shintodb/html/NA.html#method-NA-list_tables'><code>shintodb::NA$list_tables()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="make_column"><a href='../../shintodb/html/NA.html#method-NA-make_column'><code>shintodb::NA$make_column()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="postgres_now_in_local_time"><a href='../../shintodb/html/NA.html#method-NA-postgres_now_in_local_time'><code>shintodb::NA$postgres_now_in_local_time()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="query"><a href='../../shintodb/html/NA.html#method-NA-query'><code>shintodb::NA$query()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="read_table"><a href='../../shintodb/html/NA.html#method-NA-read_table'><code>shintodb::NA$read_table()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="replace_value_where"><a href='../../shintodb/html/NA.html#method-NA-replace_value_where'><code>shintodb::NA$replace_value_where()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="replace_value_where2"><a href='../../shintodb/html/NA.html#method-NA-replace_value_where2'><code>shintodb::NA$replace_value_where2()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="table_columns"><a href='../../shintodb/html/NA.html#method-NA-table_columns'><code>shintodb::NA$table_columns()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="table_info"><a href='../../shintodb/html/NA.html#method-NA-table_info'><code>shintodb::NA$table_info()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="update_where"><a href='../../shintodb/html/NA.html#method-NA-update_where'><code>shintodb::NA$update_where()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shintodb" data-topic="NA" data-id="write_table_overwrite"><a href='../../shintodb/html/NA.html#method-NA-write_table_overwrite'><code>shintodb::NA$write_table_overwrite()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-new"></a>}}
\if{latex}{\out{\hypertarget{method-NA-new}{}}}
\subsection{Method \code{new()}}{
Make new form object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$new(
  config_file = "conf/config.yml",
  what,
  schema = NULL,
  class_type = "",
  db_connection = NULL,
  filterable = FALSE,
  audit = FALSE,
  audit_table = "registrations_audit",
  def_table = "formulier_velden",
  def_columns = list(id_form = "id_formulierveld", column_field = "kolomnaam_veld",
    label_field = "label_veld", type_field = "type_veld", order_field = "volgorde_veld",
    form_section = "formulier_sectie", visible = "zichtbaar", date_deleted =
    "datum_uitgeschakeld", options = "opties", order_options = "volgorde_opties", colors
    = "kleuren", removable = "kan_worden_verwijderd", make_filter = "make_filter",
    tooltip = "tooltip"),
  data_table = "registrations",
  data_columns = list(id = "id_registratie", name = "naam_registratie", time_created =
    "invoerdatum", time_modified = "wijzigdatum", user = "user_id", status = "status",
    priority = "priority"),
  relation_table = "object_relations",
  relation_columns = list(id = "id", collector_id = "collector_id", collector_type =
    "collector_type", object_id = "object_id", object_type = "object_type", relation_id =
    "relation_id", relation_type = "relation_type", comment = "comment", status =
    "status", username = "username", timestamp = "timestamp", verwijderd = "verwijderd"),
  relation_audit_table = "object_relations_audit",
  inject = NULL,
  pool = TRUE,
  sqlite = NULL,
  mode = "default",
  default_color = "#3333cc"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config_file}}{Path to DB config}

\item{\code{schema}}{DB schema}

\item{\code{filterable}}{Boolean, TRUE if the user wants to control which registrations to filter. Requires make_filter and tooltip in the def_columns list.}

\item{\code{audit}}{If TRUE, writes edits to audit log}

\item{\code{pool}}{If TRUE, connects with dbPool}

\item{\code{sqlite}}{If path to an SQLite file, uses SQLite.}

\item{\code{default_color}}{Default color}

\item{\code{wat}}{Entry in config for DB connection}

\item{\code{table}}{Table in DB (in schema) that holds form config}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-log"></a>}}
\if{latex}{\out{\hypertarget{method-NA-log}{}}}
\subsection{Method \code{log()}}{
Log info
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$log(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-replace_value_where_multi"></a>}}
\if{latex}{\out{\hypertarget{method-NA-replace_value_where_multi}{}}}
\subsection{Method \code{replace_value_where_multi()}}{
Replace values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$replace_value_where_multi(
  table,
  replace_list,
  col_compare,
  val_compare,
  query_only = FALSE,
  quiet = FALSE,
  username = ""
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{table}}{table name}

\item{\code{replace_list}}{named list}

\item{\code{col_compare}}{name of column to compare}

\item{\code{val_compare}}{value to compare with}

\item{\code{query_only}}{if TRUE, return just the query}

\item{\code{quiet}}{if TRUE, be quiet}

\item{\code{username}}{username}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-get_by_id"></a>}}
\if{latex}{\out{\hypertarget{method-NA-get_by_id}{}}}
\subsection{Method \code{get_by_id()}}{
Get a row from the form definition by the form id.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$get_by_id(id_form)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-column_name_from_id"></a>}}
\if{latex}{\out{\hypertarget{method-NA-column_name_from_id}{}}}
\subsection{Method \code{column_name_from_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$column_name_from_id(id_form)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-get_fields_by_type"></a>}}
\if{latex}{\out{\hypertarget{method-NA-get_fields_by_type}{}}}
\subsection{Method \code{get_fields_by_type()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$get_fields_by_type(field_type)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-get_field_choices"></a>}}
\if{latex}{\out{\hypertarget{method-NA-get_field_choices}{}}}
\subsection{Method \code{get_field_choices()}}{
Get (recoded) choices for a select field
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$get_field_choices(column_field)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-make_choices"></a>}}
\if{latex}{\out{\hypertarget{method-NA-make_choices}{}}}
\subsection{Method \code{make_choices()}}{
Make choices (for selectInput) based on values and names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$make_choices(
  values_from,
  names_from = values_from,
  data = NULL,
  sort = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-choices_from_json"></a>}}
\if{latex}{\out{\hypertarget{method-NA-choices_from_json}{}}}
\subsection{Method \code{choices_from_json()}}{
Unpack a JSON field to make a named vector
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$choices_from_json(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-distinct_registration_field"></a>}}
\if{latex}{\out{\hypertarget{method-NA-distinct_registration_field}{}}}
\subsection{Method \code{distinct_registration_field()}}{
Get distinct values of a column in the registrations table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$distinct_registration_field(column_name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-rename_definition_table"></a>}}
\if{latex}{\out{\hypertarget{method-NA-rename_definition_table}{}}}
\subsection{Method \code{rename_definition_table()}}{
Rename database table to correct internal column names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$rename_definition_table(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-get_input_fields"></a>}}
\if{latex}{\out{\hypertarget{method-NA-get_input_fields}{}}}
\subsection{Method \code{get_input_fields()}}{
Get all non-deleted form input fields
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$get_input_fields(zichtbaarheid = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-get_form_fields"></a>}}
\if{latex}{\out{\hypertarget{method-NA-get_form_fields}{}}}
\subsection{Method \code{get_form_fields()}}{
Get form fields for left or right column of the form
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$get_form_fields(form_section)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-get_next_formorder_number"></a>}}
\if{latex}{\out{\hypertarget{method-NA-get_next_formorder_number}{}}}
\subsection{Method \code{get_next_formorder_number()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$get_next_formorder_number(form_section)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-add_input_field_to_form"></a>}}
\if{latex}{\out{\hypertarget{method-NA-add_input_field_to_form}{}}}
\subsection{Method \code{add_input_field_to_form()}}{
Adds an input field to a form
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$add_input_field_to_form(
  label_field,
  type_field,
  form_section,
  filterable = NULL,
  tooltip = NULL,
  column_name = NULL,
  column_2_name = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{label_field}}{Label for the field}

\item{\code{type_field}}{Type of field (options : TODO)}

\item{\code{form_section}}{Left or right}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-check_uniqueness_column_name"></a>}}
\if{latex}{\out{\hypertarget{method-NA-check_uniqueness_column_name}{}}}
\subsection{Method \code{check_uniqueness_column_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$check_uniqueness_column_name(column)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-edit_label_field"></a>}}
\if{latex}{\out{\hypertarget{method-NA-edit_label_field}{}}}
\subsection{Method \code{edit_label_field()}}{
Edit the label for an input field
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$edit_label_field(id_form, new_label)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-edit_field_type"></a>}}
\if{latex}{\out{\hypertarget{method-NA-edit_field_type}{}}}
\subsection{Method \code{edit_field_type()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$edit_field_type(id_form, new_type)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-edit_filterable_column"></a>}}
\if{latex}{\out{\hypertarget{method-NA-edit_filterable_column}{}}}
\subsection{Method \code{edit_filterable_column()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$edit_filterable_column(id_form, new_filter_boolean, new_tooltip)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-from_json"></a>}}
\if{latex}{\out{\hypertarget{method-NA-from_json}{}}}
\subsection{Method \code{from_json()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$from_json(x, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-to_json"></a>}}
\if{latex}{\out{\hypertarget{method-NA-to_json}{}}}
\subsection{Method \code{to_json()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$to_json(x, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-edit_options_field"></a>}}
\if{latex}{\out{\hypertarget{method-NA-edit_options_field}{}}}
\subsection{Method \code{edit_options_field()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$edit_options_field(id_form, new_options)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-edit_nested_column_label"></a>}}
\if{latex}{\out{\hypertarget{method-NA-edit_nested_column_label}{}}}
\subsection{Method \code{edit_nested_column_label()}}{
Label for nested column is kept in options list
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$edit_nested_column_label(id_form, opts, new_label)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-is_color"></a>}}
\if{latex}{\out{\hypertarget{method-NA-is_color}{}}}
\subsection{Method \code{is_color()}}{
Is the provided color(s) valid?
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$is_color(colors)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-edit_options_colors"></a>}}
\if{latex}{\out{\hypertarget{method-NA-edit_options_colors}{}}}
\subsection{Method \code{edit_options_colors()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$edit_options_colors(id_form, new_colors)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-amend_options_colors"></a>}}
\if{latex}{\out{\hypertarget{method-NA-amend_options_colors}{}}}
\subsection{Method \code{amend_options_colors()}}{
If new options added, make sure the color vector is amended
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$amend_options_colors(id_form, options)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-set_options_order"></a>}}
\if{latex}{\out{\hypertarget{method-NA-set_options_order}{}}}
\subsection{Method \code{set_options_order()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$set_options_order(id_form, new_order)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-prepare_nested_choice_column"></a>}}
\if{latex}{\out{\hypertarget{method-NA-prepare_nested_choice_column}{}}}
\subsection{Method \code{prepare_nested_choice_column()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$prepare_nested_choice_column(id_form, name, options)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-amend_options_order"></a>}}
\if{latex}{\out{\hypertarget{method-NA-amend_options_order}{}}}
\subsection{Method \code{amend_options_order()}}{
If new categories added, make sure the order vector is amended
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$amend_options_order(id_form, options)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-edit_formulier_setup"></a>}}
\if{latex}{\out{\hypertarget{method-NA-edit_formulier_setup}{}}}
\subsection{Method \code{edit_formulier_setup()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$edit_formulier_setup(new_setup)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-edit_zichtbaarheid_invoerveld"></a>}}
\if{latex}{\out{\hypertarget{method-NA-edit_zichtbaarheid_invoerveld}{}}}
\subsection{Method \code{edit_zichtbaarheid_invoerveld()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$edit_zichtbaarheid_invoerveld(id_formfield, new_zichtbaarheid)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-edit_verwijder_datum"></a>}}
\if{latex}{\out{\hypertarget{method-NA-edit_verwijder_datum}{}}}
\subsection{Method \code{edit_verwijder_datum()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$edit_verwijder_datum(id_formfield, new_date)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-amend_formfield_order"></a>}}
\if{latex}{\out{\hypertarget{method-NA-amend_formfield_order}{}}}
\subsection{Method \code{amend_formfield_order()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$amend_formfield_order(formside, deleted_number)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-reset_volgorde_invoerveld"></a>}}
\if{latex}{\out{\hypertarget{method-NA-reset_volgorde_invoerveld}{}}}
\subsection{Method \code{reset_volgorde_invoerveld()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$reset_volgorde_invoerveld(id_formfield, new_volgorde_nummer)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-really_delete_formfield"></a>}}
\if{latex}{\out{\hypertarget{method-NA-really_delete_formfield}{}}}
\subsection{Method \code{really_delete_formfield()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$really_delete_formfield(id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-column_type_from_field_type"></a>}}
\if{latex}{\out{\hypertarget{method-NA-column_type_from_field_type}{}}}
\subsection{Method \code{column_type_from_field_type()}}{
Database column type based on field input type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$column_type_from_field_type(type)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-prepare_data_table"></a>}}
\if{latex}{\out{\hypertarget{method-NA-prepare_data_table}{}}}
\subsection{Method \code{prepare_data_table()}}{
Make sure the necessary columns are present in the data_table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$prepare_data_table()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-rename_registrations_intern"></a>}}
\if{latex}{\out{\hypertarget{method-NA-rename_registrations_intern}{}}}
\subsection{Method \code{rename_registrations_intern()}}{
Rename registrations data (as read with $read_registrations) to
internal column names as defined in \code{data_columns} argument
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$rename_registrations_intern(data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-write_new_registration"></a>}}
\if{latex}{\out{\hypertarget{method-NA-write_new_registration}{}}}
\subsection{Method \code{write_new_registration()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$write_new_registration(data, user_id, current_reg_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-edit_registration"></a>}}
\if{latex}{\out{\hypertarget{method-NA-edit_registration}{}}}
\subsection{Method \code{edit_registration()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$edit_registration(old_data, new_data, user_id, current_reg_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-delete_registration"></a>}}
\if{latex}{\out{\hypertarget{method-NA-delete_registration}{}}}
\subsection{Method \code{delete_registration()}}{
Soft/hard/(un)delete a registration. \code{deleted} field will be set to 1.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$delete_registration(
  registration_id,
  method = c("soft", "hard", "undelete")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{registration_id}}{ID of the registration}

\item{\code{method}}{Either 'soft', 'hard' (really delete), 'undelete'}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-read_registrations"></a>}}
\if{latex}{\out{\hypertarget{method-NA-read_registrations}{}}}
\subsection{Method \code{read_registrations()}}{
Read registrations, recode select values where needed.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$read_registrations(
  recode = TRUE,
  include_deleted = TRUE,
  deleted_only = FALSE,
  lazy = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{recode}}{If TRUE (default), replaces codes with labels (for select fields)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-label_registrations_columns"></a>}}
\if{latex}{\out{\hypertarget{method-NA-label_registrations_columns}{}}}
\subsection{Method \code{label_registrations_columns()}}{
Replace column names of registrations with their labels
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$label_registrations_columns(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Read in with \code{read_registrations} or \code{get_registration_by_id}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-recode_registrations"></a>}}
\if{latex}{\out{\hypertarget{method-NA-recode_registrations}{}}}
\subsection{Method \code{recode_registrations()}}{
Recode select values in registrations data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$recode_registrations(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Dataframe (registrations data)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-get_registration_by_id"></a>}}
\if{latex}{\out{\hypertarget{method-NA-get_registration_by_id}{}}}
\subsection{Method \code{get_registration_by_id()}}{
Get a registration with a uuid
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$get_registration_by_id(id, recode = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{ID of the registration (uuid)}

\item{\code{recode}}{If TRUE, recodes integer values to their labels}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-filter_period"></a>}}
\if{latex}{\out{\hypertarget{method-NA-filter_period}{}}}
\subsection{Method \code{filter_period()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$filter_period(
  data,
  date_start = Sys.Date() - 7,
  date_end = Sys.Date()
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-read_definition"></a>}}
\if{latex}{\out{\hypertarget{method-NA-read_definition}{}}}
\subsection{Method \code{read_definition()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$read_definition(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-get_occurences"></a>}}
\if{latex}{\out{\hypertarget{method-NA-get_occurences}{}}}
\subsection{Method \code{get_occurences()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$get_occurences(table, column, record)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-get_registration_by_json"></a>}}
\if{latex}{\out{\hypertarget{method-NA-get_registration_by_json}{}}}
\subsection{Method \code{get_registration_by_json()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$get_registration_by_json(table, column, record)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-create_timeseries"></a>}}
\if{latex}{\out{\hypertarget{method-NA-create_timeseries}{}}}
\subsection{Method \code{create_timeseries()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$create_timeseries(columns = NULL, date_range = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-create_events"></a>}}
\if{latex}{\out{\hypertarget{method-NA-create_events}{}}}
\subsection{Method \code{create_events()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$create_events(timeseries, date_range = NULL, creation_only = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-get_all_relations"></a>}}
\if{latex}{\out{\hypertarget{method-NA-get_all_relations}{}}}
\subsection{Method \code{get_all_relations()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$get_all_relations()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-get_objects_for_collector"></a>}}
\if{latex}{\out{\hypertarget{method-NA-get_objects_for_collector}{}}}
\subsection{Method \code{get_objects_for_collector()}}{
get_objects_for_collector
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$get_objects_for_collector(
  collector_type = NA,
  collector_id = NA,
  relation_type = NA,
  relation_id = NA,
  object_type = NA,
  filter_verwijderd = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{collector_type}}{Type of collector}

\item{\code{collector_id}}{ID of collector}

\item{\code{relation_type}}{Description of collector -> object relation}

\item{\code{relation_id}}{ID of relation (often session ID of module)}

\item{\code{filter_verwijderd}}{when status is zero}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-get_collector_for_object"></a>}}
\if{latex}{\out{\hypertarget{method-NA-get_collector_for_object}{}}}
\subsection{Method \code{get_collector_for_object()}}{
get_collector_for_object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$get_collector_for_object(
  object_id,
  collector_type = NA,
  object_type = NA,
  relation_type = NA,
  relation_id = NA,
  filter_verwijderd = T
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{collector_type}}{Type of collector}

\item{\code{relation_type}}{Description of collector -> object relation}

\item{\code{relation_id}}{ID of relation (often session ID of module)}

\item{\code{filter_verwijderd}}{when status is zero}

\item{\code{collector_id}}{ID of collector}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-add_relation"></a>}}
\if{latex}{\out{\hypertarget{method-NA-add_relation}{}}}
\subsection{Method \code{add_relation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$add_relation(
  id = uuid::UUIDgenerate(),
  collector_id,
  collector_type,
  object_id,
  object_type,
  relation_id,
  relation_type,
  comment = NA,
  status = "1",
  verwijderd = FALSE,
  username,
  timestamp
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-postgres_now"></a>}}
\if{latex}{\out{\hypertarget{method-NA-postgres_now}{}}}
\subsection{Method \code{postgres_now()}}{
timestamp in postgres timezone
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$postgres_now()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-get_rows_since_auditstamp"></a>}}
\if{latex}{\out{\hypertarget{method-NA-get_rows_since_auditstamp}{}}}
\subsection{Method \code{get_rows_since_auditstamp()}}{
Select rows in registration audit table since some timestamp
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$get_rows_since_auditstamp(time_since)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-soft_delete_relations"></a>}}
\if{latex}{\out{\hypertarget{method-NA-soft_delete_relations}{}}}
\subsection{Method \code{soft_delete_relations()}}{
Soft delete relations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$soft_delete_relations(ids)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ids}}{ID of rows to delete}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-write_new_relations"></a>}}
\if{latex}{\out{\hypertarget{method-NA-write_new_relations}{}}}
\subsection{Method \code{write_new_relations()}}{
Append relations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$write_new_relations(data, registration_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Data to append}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NA-update_relations"></a>}}
\if{latex}{\out{\hypertarget{method-NA-update_relations}{}}}
\subsection{Method \code{update_relations()}}{
update current relation table for registration ID
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$update_relations(new_relations, registration_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_relations}}{The current state that should end up in the relations table}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-unknown-clone"></a>}}
\if{latex}{\out{\hypertarget{method-unknown-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{formClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
