% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formPageModule.R
\name{formPageUI}
\alias{formPageUI}
\alias{formPageModule}
\title{Formulier Shiny Module}
\usage{
formPageUI(
  id,
  title = "Formulier",
  icon = softui::bsicon("ui_checks"),
  header_ui = NULL,
  tag = "Opties",
  new_existing_labels = c("Nieuwe registratie invoeren",
    "Bestaande registratie bekijken")
)

formPageModule(input, output, session, .reg = NULL)
}
\description{
Formulier Shiny Module
}
