% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nieuweRegistratieModule.R
\name{nieuweRegistratieUI}
\alias{nieuweRegistratieUI}
\alias{nieuweRegistratieModule}
\title{Nieuwe registratie Shiny module}
\usage{
nieuweRegistratieUI(id)

nieuweRegistratieModule(input, output, session, .reg = NULL)
}
\description{
Nieuwe registratie Shiny module
}
