% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formModule.R
\name{formUI}
\alias{formUI}
\alias{formModule}
\title{Nieuwe registratie Shiny module}
\usage{
formUI(id, buttons = TRUE, deletable = FALSE)

formModule(
  input,
  output,
  session,
  .reg = NULL,
  ping_update = reactive(NULL),
  current_user,
  trigger = reactive(NULL),
  data = reactive(NULL),
  bucket_data = reactive(NULL),
  write_method = reactive("new"),
  use_relations = TRUE,
  use_confirmation_modal = TRUE,
  registration_description_function = function(data) NULL,
  confirm = reactive(NULL),
  cancel = reactive(NULL),
  disabled = reactive(FALSE),
  datetime_format = "\%d/\%m/\%Y \%H:\%M",
  callback_confirm = function() {
 },
  callback_cancel = function() {
 },
  message_success = "Registratie opgeslagen",
  message_error = "Er is een fout opgetreden",
  message_deleted = "Registratie verwijderd",
  shintousers_object = NULL,
  shintousers_groups = NULL,
  shintousers_ignore_groups = NULL
)
}
\description{
Nieuwe registratie Shiny module
}
