% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formModule.R
\name{formUI}
\alias{formUI}
\alias{formModule}
\title{Nieuwe registratie Shiny module}
\usage{
formUI(id, buttons = TRUE)

formModule(
  input,
  output,
  session,
  .reg = NULL,
  ping_update = reactive(NULL),
  current_user,
  data = reactive(NULL),
  write_method = reactive("new"),
  disabled = FALSE,
  confirm = reactive(NULL),
  cancel = reactive(NULL),
  callback_confirm = function() { },
  callback_cancel = function() { }
)
}
\description{
Nieuwe registratie Shiny module
}
