% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formModule.R
\name{formUI}
\alias{formUI}
\alias{formModule}
\title{Nieuwe registratie Shiny module}
\usage{
formUI(id, buttons = TRUE, deletable = FALSE)

formModule(
  input,
  output,
  session,
  .reg = NULL,
  ping_update = reactive(NULL),
  current_user,
  trigger = reactive(NULL),
  data = reactive(NULL),
  write_method = reactive("new"),
  registration_description_function = function(data) NULL,
  confirm = reactive(NULL),
  cancel = reactive(NULL),
  disabled = reactive(FALSE),
  callback_confirm = function() {
 },
  callback_cancel = function() {
 },
  message_success = "Registratie opgeslagen",
  message_error = "Er is een fout opgetreden",
  message_deleted = "Registratie verwijderd"
)
}
\description{
Nieuwe registratie Shiny module
}
